// src/store/users/users.enum.ts
enum UsersActionTypeEnum {
    USERS_LOADING = "USERS_LOADING",
    USERS_LOADED = "USERS_LOADED",
    GET_ALL_USERS = "GET_ALL_USERS",
    UPDATE_USER = "UPDATE_USER_STATUS",
    GET_USER_BY_ID = "GET_USER_BY_ID",
    DELETE_USER = "DELETE_USER",
    ADD_NEW_USER = "ADD_NEW_USER",
    DO_EMPTY_SINGLE_USER_DATA = "DO_EMPTY_SINGLE_USER_DATA"

}
  
  export default UsersActionTypeEnum;